breaking out mock into its own file so it's visible elsewhere
