+++ api/client/beacon/health/BUILD.bazel
@@ -5,9 +5,11 @@ go_library(
     srcs = [
         "health.go",
         "interfaces.go",
+        "mock.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/v5/api/client/beacon/health",
     visibility = ["//visibility:public"],
+    deps = ["@org_uber_go_mock//gomock:go_default_library"],
 )
 
 go_test(
